feat: move controls into own components
