--- client/.env.development
@@ -1,2 +1,0 @@
-NEXT_PUBLIC_GRPC_GATEWAY=http://localhost:10000
-NEXT_PUBLIC_API=http://127.0.0.1:8000/api
