+++ client/.env.local
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_API=http://127.0.0.1:8000/api
