+++ client/env.ts
@@ -1,5 +1,2 @@
 
-export const API_HOST = process.env.NEXT_PUBLIC_API
 export const NEXT_HOST = "http://localhost:3000"
-
-// export const MINIO_HOST = "http://localhost:9000"
