+++ client/lib/hooks/usePlayer.ts
@@ -1,10 +1,8 @@
 'use client'
 
 import React from "react"
-import { PlayerContext } from "../contexts/player-context"
+import {PlayerContext} from "../contexts/player-context"
 
 export default function usePlayer() {
-    const context = React.useContext(PlayerContext);
-
-    return context;
+    return React.useContext(PlayerContext);
 }
