--- client/lib/types/Player.ts
@@ -1,5 +1,0 @@
-import {Audio} from "~/lib/types/Audio";
-
-export type PlayableAudio = Audio & {
-    manifest: string;
-}
